Improve the testing of code gen
Move visualization-independent tests next to the generator.

JIRA: RAIL-4069
